+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2025-08-09
+
+### Added
+
+- Structures are now spawning and despawning by map segment (optimization!)
+- Configurable allow list for structures
+- Rope Optimizer©️ (experimental but enabled by default, mitigates the HUGE performance hit caused by large number of ropes, disable it in config if ropes become unusable)
+
+### Changed
+- Fixed a bug that caused many structures from a several previous runs to appear for reconnecting clients.
+- Default limit for online structures bumped to 40
+- Server-side limit for online structures will be bumped to 300 following this mod update
+- Slightly better steam ticket handling so the client can survive the server restart
+
 ## [0.8.1] - 2025-08-03
 
 ### Added